refactor(code_gen): extract methods/ with string receiver split
